requirements.txt
Flask==3.0.3
flask-socketio==5.3.6
numpy==1.26.4
deepspeech==0.9.3
gtts==2.5.1
ffmpeg-python==0.2.0
openai==1.35.15
argparse==1.4.0
